+++ client.js
@@ -2,11 +2,12 @@
 // Imports/Dependencies
 const WebSocket = require('ws');
 const exec = require('child_process').exec;
+const ENV = process.env;
 
-const ws = new WebSocket('ws://192.168.10.242:9000');
+const ws = new WebSocket(ENV.WS_SERVER);
 
 ws.on('open', () => {
-  ws.send('RX4 connected');
+  ws.send(`{ENV.NAME} has connected.`);
 });
 
 const convertNs = (t, ns) => {
@@ -22,10 +23,12 @@ const convertNs = (t, ns) => {
 const Now = () => process.hrtime.bigint();
 
 const sample = (cmd) => new Promise(r => {
+  const start = Now();
   console.log('scanning...');
   exec(cmd, (err, out) => {
-    if (err) return r({ err, out });
-    r({ err, out }); 
+    const t = parseInt(Now() - start);
+    if (err) return r({ err, out, t });
+    r({ err, out, t }); 
   });
 });
 
@@ -38,7 +41,7 @@ ws.on('message', async (msg) => {
   const mark = (d.execT - offset) * 1000000;
   // const delay = await new Promise(r => setTimeout(r, d.execT - t));
   let end = now;;
-  for (;mark > c;) {
+  while (mark > c) {
     if (now > end) {
       c += now - end;
       end = now;
@@ -48,17 +51,21 @@ ws.on('message', async (msg) => {
   }
 
   console.log('off: ', offset, 'now: ', now, ' end: ', end, ' st: ', d.serverTime, d.execT, 'c: ', c, '\nmark: ', mark);
+  const calcLag = parseInt(c) - mark;
   const startedAt = new Date().getTime();
   const cmd = 'rtl_power -f 153084000:153304000:0.8k -g 35 -i 0 -e -1 2>&1';
   const raw = await sample(cmd);
-  const { err, out } = raw;
-  const samplingTime = convertNs('micro', parseInt(Now() - now));
+  const { err, out, t} = raw;
 
   const res = {
-    name: 'RX 4',
+    name: ENV.NAME,
     msg: 'Sampling done!',
+    calcLag,
     startTime: unixTS, 
-    samplingTime,
+    samplingTime: {
+      unit: 'ns',
+      t
+    },
     startedAt,
     data: err ? null : out,
     error: err ? 'error...' : null,