+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 [0.0.1]
 ### Updates
+- execution counter
 - hrtime.bigint() to calculate execution time
 
 [0.0.0]