+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 [0.0.1]
 ### Updates
+- server variables [ NAME, WS_SERVER ] in .bashrc file
 - execution counter
 - hrtime.bigint() to calculate execution time