+++ changelog.md
@@ -1,7 +1,10 @@
 # Change Log
+[0.0.2]
+### Updates
+- Updated client to use socket.io events
+
 [0.0.1]
 ### Updates
-- 
 - server variables [ NAME, WS_SERVER ] in .bashrc file
 - execution counter
 - hrtime.bigint() to calculate execution time